pyproject.toml
[tool.poetry]
name = "nonebot2-plugins"
version = "0.1.0"
description = ""
authors = ["Chendihe4975 <Chendihe4975@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-alpha.15"
nonebot-adapter-cqhttp = "^2.0.0-alpha.15"
aiohttp = "^3.7.4"
peewee = "^3.14.4"
ujson = "^4.1.0"
Pillow = "^8.3.1"
nonebot-plugin-apscheduler = "^0.1.2"
aiofiles = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
